Prepare for a more exhaustive test-suite, using JSON

+++ src/main/java/fr/mieuxvoter/mj/TallyInterface.java
@@ -1,11 +1,13 @@
 package fr.mieuxvoter.mj;
 
+import java.math.BigInteger;
+
 public interface TallyInterface {
-	
+
 	public ProposalTallyInterface[] getProposalsTallies();
-	
-	public Long getAmountOfJudges();
-	
+
+	public BigInteger getAmountOfJudges();
+
 	public Integer getAmountOfProposals();
-	
+
 }
